Add proposal for the compact code model
Add a new file containing the draft of a new code model whose purpose is to
support systems with sparse memory maps.  In particular when the device
containing the code is different from the one containing the data.
Additionally, I/O buffers are also supported by this code model, through the
use of linker scripts to locate them in the memory map.

Yet, though a large amount of static data is also possible, though limited
to 2GiB when the scope is local, it is not the motivation of this design and
is deemed unlikely to be widely used.